remove node_module volume
These get run as part of the dockerfile and so they're regenerated anyways. I found that pull in this as a volume overwrote writing my new dependencies after I've run the app a few times and so it prevented my dependencies from taking shape. I think we can remove it without harm.